chore(server): quick js to ts #5 - remaining stragglers
